requirements.txt
accelerate==1.8.1
altair==5.5.0
attrs==25.3.0
blinker==1.9.0
cachetools==6.1.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
et_xmlfile==2.0.0
filelock==3.18.0
Flask==3.1.1
flask-cors==6.0.1
Flask-SQLAlchemy==3.1.1
fsspec==2025.5.1
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.2.3
huggingface-hub==0.33.2
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
MarkupSafe==3.0.2
mpmath==1.3.0
narwhals==1.45.0
networkx==3.5
nltk==3.9.1
numpy==2.3.1
openpyxl==3.1.5
packaging==25.0
pandas==2.3.0
pillow==11.3.0
pluggy==1.6.0
protobuf==6.31.1
psutil==7.0.0
pyarrow==20.0.0
pydeck==0.9.1
Pygments==2.19.2
pytest==8.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
rpds-py==0.26.0
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.16.0
sentencepiece==0.2.0
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
SQLAlchemy==2.0.41
streamlit==1.46.1
sympy==1.14.0
tenacity==9.1.2
textblob==0.19.0
threadpoolctl==3.6.0
tokenizers==0.21.2
toml==0.10.2
torch==2.7.1
tornado==6.5.1
tqdm==4.67.1
transformers==4.53.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
Werkzeug==3.1.3
wheel==0.43.0
xlsxwriter==3.2.5 